backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Limiter
Flask-Caching
python-dotenv
psycopg2-binary
gunicorn
marshmallow
qrcode[pil]
flask-socketio
eventlet
